Update Go package management for modules
As of 1.18, `go install` is the only supported way to install binaries
(replacing `go get -u`). As part of this change, a single invocation of
`go install` can only install packages from a single module. The
simplest way to accommodate this is to just put the Go packages I'm
using into the Makefile.